Version1.0.14

+++ include/planner/planner.h
@@ -4,10 +4,12 @@
 #include<iostream>
 #include<math.h>
 #include<vector>
+#include<list>
 using namespace std;
 struct Point{
     int x,y,z;
     int F,G,H;
+    
     Point* parent;
     Point(int _x,int _y,int _z):x(_x),y(_y),z(_z),F(0),G(0),H(0),parent(NULL){};
     
@@ -16,7 +18,7 @@ struct Point{
     public:
         std::vector<std::vector<std::vector<int>>>map;
         std::vector<Point*> path;
-        virtual std::list<Point*>* getPath(Point *start_point,Point *end_point,double &time,int &distance){};
+        virtual std::list<Point*>* getPath(Point *start_point,Point *end_point,double &time,double &distance){};
 };
